add utils.cpp
